Fix OpenGL 3 check Fix implicit cast warnings

+++ ouzel/FileSystem.cpp
@@ -146,7 +146,7 @@ namespace ouzel
 
         file.seekg(0, std::ios_base::end);
         std::streampos fileSize = file.tellg();
-        data.resize(fileSize);
+        data.resize(static_cast<size_t>(fileSize));
 
         file.seekg(0, std::ios_base::beg);
         file.read(reinterpret_cast<char*>(&data[0]), fileSize);
